Split off preprocessor MeshEntity/Element from mesh
While they are conceptually intertwined, having them in separate files
makes it easier to work with and have an overview of each class.
