+++ applications/Preprocessor/lib/mesh_impl.h
@@ -36,240 +36,10 @@
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "mesh.h"
-
 #include "ElementShapes.h"
 
 namespace Preprocessor {
 
-template <std::size_t entityDimension, std::size_t meshDimension>
-Element<meshDimension>& MeshEntity<entityDimension, meshDimension>::getElement(
-    EntityLId i) {
-    return mesh->getElement(elementIDs[i]);
-};
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-const Element<meshDimension>&
-    MeshEntity<entityDimension, meshDimension>::getElement(EntityLId i) const {
-    return mesh->getElement(elementIDs[i]);
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-EntityLId MeshEntity<entityDimension, meshDimension>::getElementIndex(
-    const Element<meshDimension>& element) const {
-    for (std::size_t i = 0; i < elementIDs.size(); ++i) {
-        if (getElement(i) == element) {
-            return i;
-        }
-    }
-    logger(ERROR, "Element not found");
-    return 0;
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-std::size_t MeshEntity<entityDimension, meshDimension>::getNumberOfElements()
-    const {
-    return elementIDs.size();
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-EntityLId MeshEntity<entityDimension, meshDimension>::getLocalIndex(
-    EntityLId i) const {
-    return localIDs[i];
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-EntityLId MeshEntity<entityDimension, meshDimension>::getLocalIndex(
-    const Element<meshDimension>& element) const {
-    for (std::size_t i = 0; i < elementIDs.size(); ++i) {
-        if (getElement(i).getGlobalIndex() == element.getGlobalIndex()) {
-            logger.assert_debug(getElement(i) == element,
-                                "Two different elements got the same index");
-            return localIDs[i];
-        }
-    }
-    logger(ERROR, "Element not found");
-    return 0;
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-EntityGId MeshEntity<entityDimension, meshDimension>::getGlobalIndex() const {
-    return entityID;
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-template <int d>
-std::vector<MeshEntity<(d < 0 ? d + meshDimension : d), meshDimension>>
-    MeshEntity<entityDimension, meshDimension>::getIncidenceList() const {
-    auto indices = getIncidenceListAsIndices<d>();
-    std::vector<MeshEntity<(d < 0 ? d + meshDimension : d), meshDimension>>
-        result(indices.size());
-    for (std::size_t i = 0; i < indices.size(); ++i) {
-        result[i] = mesh->template getEntity<(d < 0 ? d + meshDimension : d)>(
-            indices[i]);
-    }
-    return result;
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-template <int d>
-std::vector<EntityGId> MeshEntity<
-    entityDimension, meshDimension>::getIncidenceListAsIndices() const {
-    static_assert(d + entityDimension >= 0,
-                  "The codimension you are interested in is too high for the "
-                  "dimension of this object");
-    if (d < 0) {
-        return getIncidenceListAsIndices<(d < 0 ? d + meshDimension : d)>();
-    }
-    if (d == entityDimension) {
-        return {getGlobalIndex()};
-    }
-    if (d == meshDimension) {
-        return std::vector<EntityGId>(elementIDs.begin(), elementIDs.end());
-    } else if (d < entityDimension) {
-        // easy case: all adjacent entities of this entityDimension are adjacent
-        // to all adjacent elements
-        return getElement(0).template getIncidenceListAsIndices<d>(*this);
-    } else {
-        std::vector<EntityGId> result;
-        for (auto elementID : elementIDs) {
-            const auto& element = mesh->getElement(elementID);
-            std::vector<EntityGId> newEntries =
-                element.template getIncidenceListAsIndices<d>(*this);
-            result.insert(result.end(), newEntries.begin(), newEntries.end());
-        }
-        std::sort(result.begin(), result.end());
-        result.erase(std::unique(result.begin(), result.end()), result.end());
-        return result;
-    }
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-void MeshEntity<entityDimension, meshDimension>::addElement(
-    EntityGId elementID, EntityLId localEntityIndex) {
-    elementIDs.push_back(elementID);
-    localIDs.push_back(localEntityIndex);
-}
-
-template <std::size_t entityDimension, std::size_t meshDimension>
-bool MeshEntity<entityDimension, meshDimension>::operator==(
-    const MeshEntity& other) const {
-    return mesh == other.mesh && entityID == other.entityID &&
-           elementIDs == other.elementIDs && localIDs == other.localIDs;
-}
-
-template <std::size_t dimension>
-bool Element<dimension>::operator==(const Element& other) const {
-    return static_cast<const MeshEntity<dimension, dimension>&>(*this) ==
-               static_cast<const MeshEntity<dimension, dimension>&>(other) &&
-           other.globalCoordinateIndices == globalCoordinateIndices &&
-           other.incidenceLists == incidenceLists;
-}
-
-template <std::size_t dimension>
-LinearAlgebra::SmallVector<dimension> Element<dimension>::getCoordinate(
-    EntityLId localIndex) const {
-    return this->mesh->getCoordinate(globalCoordinateIndices[localIndex.id]);
-}
-
-template <std::size_t dimension>
-CoordId Element<dimension>::getCoordinateIndex(EntityLId localIndex) const {
-    return globalCoordinateIndices[localIndex.id];
-}
-
-template <std::size_t dimension>
-std::vector<LinearAlgebra::SmallVector<dimension>>
-    Element<dimension>::getCoordinatesList() const {
-    std::vector<LinearAlgebra::SmallVector<dimension>> result;
-    result.reserve(referenceGeometry->getNumberOfNodes());
-    for (std::size_t i = 0; i < referenceGeometry->getNumberOfNodes(); ++i) {
-        result.push_back(getCoordinate(i));
-    }
-    return result;
-}
-
-template <std::size_t dimension>
-void Element<dimension>::setNodeCoordinate(
-    EntityLId localIndex, LinearAlgebra::SmallVector<dimension> newCoordinate) {
-    CoordId globalIndex = globalCoordinateIndices[localIndex.id];
-    this->mesh->updateCoordinate(globalIndex, newCoordinate);
-}
-
-template <std::size_t dimension>
-void Element<dimension>::setNode(EntityLId localIndex, EntityGId globalIndex,
-                                 CoordId coordinateIndex) {
-    incidenceLists[0][localIndex.id] = globalIndex;
-    globalCoordinateIndices[localIndex.id] = coordinateIndex;
-}
-
-template <std::size_t dimension>
-template <std::size_t d>
-std::enable_if_t<(d > 0)> Element<dimension>::setEntity(EntityLId localIndex,
-                                                        EntityGId globalIndex) {
-    incidenceLists[d][localIndex] = globalIndex;
-}
-
-template <std::size_t dimension>
-template <int d, std::size_t entityDimension>
-std::vector<MeshEntity<(d < 0 ? d + dimension : d), dimension>>
-    Element<dimension>::getIncidenceList(
-        const MeshEntity<entityDimension, dimension>& entity) const {
-    auto indices = getIncidenceListAsIndices<d>(entity);
-    std::vector<MeshEntity<(d < 0 ? d + dimension : d), dimension>> result(
-        indices.size());
-    for (std::size_t i = 0; i < indices.size(); ++i) {
-        result[i] = this->mesh->template getEntity<(d < 0 ? d + dimension : d)>(
-            indices[i]);
-    }
-}
-
-template <std::size_t dimension>
-template <int d, std::size_t entityDimension>
-std::vector<EntityGId> Element<dimension>::getIncidenceListAsIndices(
-    const MeshEntity<entityDimension, dimension>& entity) const {
-    auto localIndices = getLocalIncidenceListAsIndices<d>(entity);
-    std::vector<EntityGId> globalIndices(localIndices.size());
-    for (std::size_t i = 0; i < globalIndices.size(); ++i) {
-        globalIndices[i] =
-            incidenceLists[(d < 0 ? d + dimension : d)][localIndices[i]];
-    }
-    return globalIndices;
-}
-
-template <std::size_t dimension>
-template <int d, std::size_t entityDimension>
-std::vector<EntityLId> Element<dimension>::getLocalIncidenceListAsIndices(
-    const MeshEntity<entityDimension, dimension>& entity) const {
-    static_assert(d + dimension >= 0,
-                  "The requested codimension is too high for the dimension of "
-                  "this element");
-    constexpr std::size_t actualDimension = (d < 0 ? d + dimension : d);
-    if (incidenceLists[actualDimension].size() <
-        referenceGeometry->template getNumberOfEntities<actualDimension>())
-        return {};
-    std::vector<std::size_t> result =
-        referenceGeometry
-            ->template getAdjacentEntities<entityDimension, actualDimension>(
-                entity.getLocalIndex(*this));
-    std::vector<EntityLId> realResult(result.size());
-    for (std::size_t i = 0; i < result.size(); ++i) {
-        realResult[i] = EntityLId(result[i]);
-    }
-    return realResult;
-}
-
-template <std::size_t dimension>
-void Element<dimension>::addNode(EntityGId globalNodeIndex,
-                                 CoordId coordinateIndex) {
-    incidenceLists[0].push_back(globalNodeIndex);
-    globalCoordinateIndices.push_back(coordinateIndex);
-}
-
-template <std::size_t dimension>
-template <std::size_t d>
-std::enable_if_t<(d > 0)> Element<dimension>::addEntity(EntityGId globalIndex) {
-    incidenceLists[d].push_back(globalIndex);
-}
 
 template <std::size_t dimension>
 std::vector<Element<dimension>>& Mesh<dimension>::getElements() {
@@ -323,7 +93,7 @@ std::vector<MeshEntity<(entityDimension < 0 ? entityDimension + dimension
                   "this element");
     constexpr std::size_t actualDimension =
         (entityDimension < 0 ? entityDimension + dimension : entityDimension);
-    return otherEntities.template get<actualDimension>();
+    return meshEntities.template get<actualDimension>();
 };
 
 template <std::size_t dimension>
@@ -337,7 +107,7 @@ const std::vector<MeshEntity<(entityDimension < 0 ? entityDimension + dimension
                   "this element");
     constexpr std::size_t actualDimension =
         (entityDimension < 0 ? entityDimension + dimension : entityDimension);
-    return otherEntities.template get<actualDimension>();
+    return meshEntities.template get<actualDimension>();
 }
 
 template <std::size_t dimension>
@@ -354,10 +124,10 @@ const std::vector<typename Mesh<dimension>::coordinateData>&
 
 template <std::size_t dimension>
 void Mesh<dimension>::setNumberOfNodes(std::size_t number) {
-    if (number < otherEntities.template get<0>().size()) {
-        otherEntities.template get<0>().resize(number);
+    if (number < meshEntities.template get<0>().size()) {
+        meshEntities.template get<0>().resize(number);
     } else
-        addNodes(number - otherEntities.template get<0>().size());
+        addNodes(number - meshEntities.template get<0>().size());
 }
 
 template <std::size_t dimension>
@@ -390,14 +160,32 @@ void Mesh<dimension>::addElement(std::vector<CoordId> nodeCoordinateIDs,
         newElement.addNode(nodeID, coordinateID);
     }
     elementsList.push_back(newElement);
-    otherEntities.template get<dimension>().push_back(newElement);
+    meshEntities.template get<dimension>().push_back(newElement);
     for (std::size_t i = 0; i < nodeCoordinateIDs.size(); ++i) {
         auto coordinateID = nodeCoordinateIDs[i];
         EntityGId nodeID = coordinates[coordinateID.id].nodeIndex;
-        otherEntities.template get<0>()[nodeID.id].addElement(elementID, i);
+        meshEntities.template get<0>()[nodeID.id].addElement(elementID, i);
     }
     fixElement(elementsList.back(), tag<dimension - 1>{});
 }
+
+template<std::size_t dimension>
+bool Mesh<dimension>::isValid() const {
+    logger(DEBUG, "The mesh has % elements", getNumberOfElements());
+    for (auto element : getElements()) {
+        if (getElement(element.getGlobalIndex()) != element) {
+            logger(ERROR,
+                   "The index for element % has been set incorrectly",
+                   element.getGlobalIndex().id);
+            return false;
+        }
+        if (!checkBoundingEntities(element, tag<dimension - 1>{})) {
+            return false;
+        }
+    }
+    return checkEntities(tag<dimension>{});
+}
+
 
 template <std::size_t dimension>
 void Mesh<dimension>::fixConnectivity() {
@@ -456,7 +244,7 @@ void Mesh<dimension>::fixEntity(Element<dimension>& element,
     EntityGId entityIndex =
         (candidates.size() == 1 ? candidates[0] : newEntity<d>());
     element.template addEntity<d>(entityIndex);
-    otherEntities.template get<d>()[entityIndex.id].addElement(
+    meshEntities.template get<d>()[entityIndex.id].addElement(
         element.getGlobalIndex(), index);
 }
